moving file path into option

+++ lib/kube_secrets_encode.rb
@@ -8,13 +8,22 @@ module KubeSecretsEncode
   class Command
     include Executable
 
+    # filename of file being encoded or decoded
+    def file=(filename)
+      @file = filename
+    end
+    def file
+      @file
+    end
+
+    # enables write after decode or encode
     def yes=(bool)
       @yes = bool
     end
-
     def yes?
       @yes
     end
+    # switch from default encode to decode mode
     def decode=(bool)
       @decode = bool
     end
@@ -29,29 +38,38 @@ module KubeSecretsEncode
     end
     alias :h? :help?
 
-    # Say hello.
-    def call(file)
+    def self.usage_name
+      "kube_secrets"
+    end
+
+    # kube_secrets enables you to encode and decode the base64 encoding of secrets in place within your config files making it easier to modify them.
+    def call()
       documents = []
-      YAML.load_stream(File.open(file)) do |document|
-        if document["kind"] == "Secret"
-          if document["data"]
-            document["data"].each do |k,v|
-              if decode?
-                document["data"][k] = Base64.strict_decode64(v)
-              else
-                document["data"][k] = Base64.strict_encode64(v)
+      if file && File.exists?(file)
+        YAML.load_stream(File.open(file)) do |document|
+          if document["kind"] == "Secret"
+            if document["data"]
+              document["data"].each do |k,v|
+                if decode?
+                  document["data"][k] = Base64.strict_decode64(v)
+                else
+                  document["data"][k] = Base64.strict_encode64(v)
+                end
               end
             end
           end
+          documents << document
         end
-        documents << document
-      end
-      yaml = documents.map{|d| d.to_yaml}.join
+        yaml = documents.map{|d| d.to_yaml}.join
 
-      puts yaml
-      if yes?
-        File.open(file, 'w') { |file| file.write(yaml) }
+        puts yaml
+        if yes?
+          File.open(file, 'w') { |file| file.write(yaml) }
+        end
+      else
+        cli.show_help
       end
+      exit
     end
   end
 end
